cloud_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery
flask
google-cloud-pubsub